% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clariostarparser-package.R
\docType{package}
\name{clariostarparser-package}
\alias{clariostarparser}
\alias{clariostarparser-package}
\title{clariostarparser: Parse BMG CLARIOstar Plate Reader Excel Exports}
\description{
Parses Excel exports from BMG CLARIOstar plate readers into tidy
long-format data frames, supporting both the "Table All Data points"
and "Microplate End point" export formats. See \code{\link[=read_clariostar]{read_clariostar()}} for
the main entry point.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jackspepper/tools}
  \item Report bugs at \url{https://github.com/jackspepper/tools/issues}
}

}
\author{
\strong{Maintainer}: Jack Pepper \email{jack.pepper@thekids.org.au}

Authors:
\itemize{
  \item Jack Pepper \email{jack.pepper@thekids.org.au}
}

}
\keyword{internal}
